Add slider for graph size
Instead of having to type a grid input and deal with user error a slider makes the project much more user friendly and easy to use. Implemented using change event listeners to process the graph size but a constant event listener for the size span above to make choosing much more predictable.