Fix left click toggle bug and individual cell paint bug
Users could previouly drag there mouse while holding click off of the main container and return to hover like drawing instead of having to hold down to click. To fix, I changed the event listener for mousedown / mouseup to the body instead of just the container used to draw in. Changed the mouseenter event listener to draw to a mouseover event listener instead so drawing in indiviidual boxes is much easier.